+++ admin/tabs/general/users/UserPanel.js
@@ -9,7 +9,7 @@ import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
 import {filler} from '@xh/hoist/cmp/layout';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
-import {storeFilterField} from '@xh/hoist/desktop/cmp/store';
+import {storeFilterField} from '@xh/hoist/cmp/store';
 import {switchInput} from '@xh/hoist/desktop/cmp/input';
 import {exportButton} from '@xh/hoist/desktop/cmp/button';