+++ CHANGELOG.md
@@ -6,7 +6,6 @@
 
 * Added a `showCounts` option to show the number of items on each side of a LeftRightChooser.
 * Added an `useOnResize` hook, which runs a function when a component is resized.
-* Added `StoreFilterField` to the mobile toolkit. Its API is identical to the desktop version.
 
 ### 💥 Breaking Changes
 
@@ -14,7 +13,8 @@
 parameter will allow context menus to be specifed with a simple array in addition to the function
 specification currently supported.
 * The GridModel `defaultContextMenuTokens` array has been renamed `defaultContextMenu`.
-* Charts have been moved from `desktop/cmp/charts` to `cmp/charts`.
+* `Chart` and `ChartModel` have been moved from `desktop/cmp/charts` to `cmp/charts`.
+* `StoreFilterField` have been moved from `desktop/cmp/store` to `cmp/store`.
 
 ### 🐞 Bug Fixes